Avoid overhead of threading when search space is small
Don't spawn threads if the search space is small (ie. < 1,000 paths).
Additionally, instead of spawning N threads, spawn N-1 and use the main
thread as the last worker. This allows us to keep the overhead as low as
possible, and also makes things a bit DRY-er.

This brings us back to the unmitigated win territory:

Baseline (no threading):
                          user     system      total        real
  pathological        0.000000   0.000000   0.000000 (  0.004329)
  command-t           0.420000   0.000000   0.420000 (  0.416208)
  chromium (subset)   1.070000   0.000000   1.070000 (  1.071452)
  chromium (whole)    3.680000   0.010000   3.690000 (  3.683207)

With threading (unconditional):
                          user     system      total        real
  pathological        0.010000   0.030000   0.040000 (  0.018285)
  command-t           0.980000   1.850000   2.830000 (  1.391637)
  chromium (subset)   1.480000   0.020000   1.500000 (  0.551915)
  chromium (whole)    5.150000   0.020000   5.170000 (  1.861232)

With threading enabled only for large search spaces:
                          user     system      total        real
  pathological        0.000000   0.000000   0.000000 (  0.004627)
  command-t           0.420000   0.000000   0.420000 (  0.423317)
  chromium (subset)   1.470000   0.020000   1.490000 (  0.551474)
  chromium (whole)    5.050000   0.010000   5.060000 (  1.866479)

It's possible that the threshold should/could be tweaked some more, but
for now 1,000 seems like a reasonable cut-off.

Signed-off-by: Wincent Colaiuta <win@wincent.com>