Dynamically detect presence of pthread.h
This should hopefully do the right thing, even on Windows.
HAVE_PTHREAD_H should not be set if the header can't be found, so the
threading code should degrade to the main-thread-only version.

Note there is some weirdness here, at least on OS X: as soon as you
include ruby.h, you'll end up including:

  /System/Library/Frameworks/Ruby.framework/Versions/2.0/Headers/ruby/config.h

which itself sets HAVE_PTHREAD_H. This is why I moved the:

  #ifdef HAVE_PTHREAD_H
  #include <pthread.h>
  #endif

guard clause down, until after the inclusion of ruby.h (ie. to avoid the
pthread.h header not being included if mkmf couldn't find it, but then
sections further on down the file from being activated thanks to
config.h). It turns out that this works in any case, just with compiler
warnings about use of undeclared functions, but it's better to be safe
than sorry.

Signed-off-by: Wincent Colaiuta <win@wincent.com>