Doc: apply other updates for OS X
Since these sections of the docs were written, we've had several OS X
releases and a few things have changed (not least of which is that Apple
started calling the OS "OS X" instead of "Mac OS X").

This commit updates the names, and corrects the claim that the system
Vim doesn't support Ruby (it actually does, at least on Mavericks; I am
not actually sure when that started to be true).

Finally, the principal distribution mode for Xcode has changed, so
that's noted too.

Signed-off-by: Wincent Colaiuta <win@wincent.com>