Defer processor count detection until runtime
As suggested here:

  https://wincent.com/issues/2133

Copying the text of the linked issue:

  You assume, that the Ruby extension is build on the same machine,
  where the code is executed. This assumption is invalid for Fedora.
  When I package command-t for use with Fedora, the extension is built
  on Koji builder. There might be powerful CPUs with plenty of cores,
  but it says nothing about my machine.

I've moved the processor count detection into a runtime method on a new
CommandT::Util module (seeing as there isn't any other obvious place to
put it right now). The count is memoized to avoid exposure.

In the C code that consumes the count (the Matcher#sorted_matches_for
method), I do do a `nil` check because there are a few callers in places
like the test suite that shouldn't have to bother with passing it in,
but I don't worry about more robust sanity checking, leaving that up to
the CommandT::Util method instead.

Signed-off-by: Wincent Colaiuta <win@wincent.com>