Add `find`-based scanner
This is another enhancement based off of a patch by Yiding Jia. It
allows you to choose to use a `find`-based scanner, which may be
substantially faster than doing it all in Ruby.

I have to benchmark this; it's possible that by respecting 'wildignore'
the speed benefits of shelling out to a pure-C tool will be negated
(Yiding's original patch did no filtering). At least as a starting
point, I want this to be as close to feature-parity as possible with the
Ruby scanner, which means that it respects g:CommandTMaxDepth,
g:CommandTMaxFiles, g:CommandTScanDotDirectories, and
g:CommandTWildIgnore.

To try out the scanner, use:

  :let g:CommandTFileScanner='find'

Signed-off-by: Wincent Colaiuta <win@wincent.com>