Add some more DB operations to complete benchmarks
Here are the results from a sample run of an in-memory database
with 400,000 rows (ie. 400k INSERTs, 100k UPDATEs, a SELECT of 400k
rows, and 100k DELETEs).

  Timings:
    Connect 0.002 seconds
    PRAGMA 0.001 seconds
    CREATE TABLE 0.001 seconds
    CREATE INDEX 0.001 seconds
    INSERT 21.168 seconds
    UPDATE 1 hour, 46 minutes, 58 seconds
    SELECT 0.953 seconds
    DELETE 1 hour, 27 minutes, 8 seconds
  Total: 3 hours, 14 minutes, 29 seconds

It is pretty slow... I'm particularly surprised at how slow UPDATEs are
compared to INSERTs. Likewise, the DELETE performance is surprising.

This might actually be OK as large parts of the tree are unlikely to be
changing at a time.

Signed-off-by: Wincent Colaiuta <win@wincent.com>