Speed up WatchmanFileScanner by an order of magnitude
It was already pretty fast (about 10 seconds on a large hierarchy that
I've been testing it with), but this commit drops that to about 1
second.

I haven't profiled this to know exactly where the time was/is being
spent, but this change massively reduces the payload that Watchman has
to send over the socket (ie. just the name and none of the metadata, and
only info for files rather than all tracked paths), which means quicker
delivery of the payload, quicker parsing, and less memory pressure from
intermediate objects.

The way it was working before actually had one unintended behavior,
which was to expose directories in the listing as well. This was kind of
interesting, as Vim could "open" those directories for you. But it was
also inconsistent with all the other scanners, and on a very large
project, those directories could often get in the way of finding files,
which is what this plug-in is all about.

Signed-off-by: Wincent Colaiuta <win@wincent.com>