Fix edge case in "Find" scanner with relative path
When Command-T was invoked with a path argument (eg. `:CommandT foo`),
the paths in the listing would show up as "foo/thing" etc, and when
selected would open as "foo/foo/thing" etc.

This commit fixes that oversight, and brings the "Find" scanner back
into line with the behavior of the default Ruby scanner. The Watchman
scanner was unaffected by this bug, except in cases where it fell back
to using the "Find" scanner as a backup.

Signed-off-by: Wincent Colaiuta <win@wincent.com>