Remove clearmatches() call
This was causing problems with the vim-css-color plug-in, blowing away
the matches that it was setting up. When I added this a long time ago in
commit d06f51f, I thought it only affected the current window, but I was
evidently wrong about that. (I'm guessing I probably didn't notice
because I likely had some autocmd that was restoring the settings any
time I opened a new file.)

The original use case was to prevent the rules that I had set up to
highlight long lines from applying in the Command-T match listing. There
are better ways to do that kind of highlighting anyway (like the
`'colorcolumn'` setting), so let's just remove the problematic call.

See:

    https://github.com/ap/vim-css-color/issues/31

Signed-off-by: Wincent Colaiuta <win@wincent.com>