Move RubyFileScanner specs into subdirectory
We had 4 remaining spec failures in this file; this commit tries to
stick as close to being a straight move as possible to make it clear
what's happening, but just moving the code makes the specs pass.

The issue is that we recently factored some of the functionality of the
FileScanner class into a RubyFileScanner subclass, so the specs for the
FileScanner were seeking to verify behavior which was no longer present
in the class.

In this commit we split the specs out just as we split the FileScanner
and RubyFileScanner functionality out. It moves the code into another
file and changes the references to the FileScanner class to
RubyFileScanner.

Signed-off-by: Wincent Colaiuta <win@wincent.com>