Further abstract local vs global distinction inside Settings class
Commit 514194c made the Settings class a little smarter. We can make the
MatchWindow class (the sole client of the Settings class) simpler by
burying the details about the distinction between local and global
settings inside of the Settings class.

Specifically, the Settings class knows which settings are global and
which are local, so the calling MatchWindow class just tells it to apply
settings and lets the Settings class worry about exactly how that should
happen.

This is a net increase in the number of lines, but worth it because the
Settings class will be infrequently touched, and the more
actively-modified MatchWindow class is simpler for it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>