Convert a couple more settings manipulations in the MatchWindow class
I originally didn't include these in the recent refactorings of the
Settings class as they have a somewhat different lifecycle than the
other settings changes made by the MatchWindow class (ie. most of the
changes are made as a batch on initial display and restored at the end,
but the `#lock` and `#unlock` calls can be made at distinct moments as
the result of other methods being called).

Nevertheless, they are `:setlocal` calls, so routing them through the
Settings class is not going to disturb its internal state anyway. In the
interests of reducing noise in the MatchWindow class and getting maximum
re-use out of the Settings abstraction, this commit swaps them over to
use the "DSL".

Signed-off-by: Wincent Colaiuta <win@wincent.com>