Fix compilation errors on 1.8
1.8 has an st.h header which defines ST_CONTINUE (et al), which is
apparently not included by default.

1.9 has a ruby/st.h header, included by default, and also a st.h header
for backwards compatibility. The latter works but issues a warning to
the console.

Let's check for the presence of the headers and use the appropriate one.
Note that ruby/st.h has the standard include guards in place, so we need
not worry about inadvertently including things twice.

Reported here:

  https://github.com/wincent/Command-T/issues/67#issuecomment-38815959

Leads to:

  watchman.c:188:12: error: use of undeclared identifier 'ST_CONTINUE'

Signed-off-by: Wincent Colaiuta <win@wincent.com>