Get rid of $scanner global variable
Move the concern of tracking MRU buffers out of the MruBufferScanner and
into a dedicated module, thus eliminating the $scanner global variable.

It's true that the new (singleton) MRU module is a global in disguise,
but at least it is a namespaced global which means that we're not
polluting the Ruby environment that other plug-ins might be operating in
with additional variables.

As a bonus, this allows us to keep some noise out of the scanner, and
encapsulate it in a well-contained module with a single purpose.

Also, this means that we track seen buffers from the very beginning, as
soon as the plug-in is loaded. Previously, tracking would only start as
a side effect of a call to the `#list_matches` method in the controller
(because it calls `#buffer_finder`, which instantiates the
MruBufferFinder and memoizes it).

Signed-off-by: Wincent Colaiuta <win@wincent.com>