Only move MRU results if appropriate
We had a bug where we would unconditionally move the first search result
to the end, thinking that it was the current buffer.

This is not necessarily always for the case. For example, if I've
visited:

  foo1
  foo2
  bar
  baz

in that order, then it's correct to move baz to the end of the list. But
if I filter my query by typing "f", only "foo1" and "foo2" will be
shown, and it would be incorrect to attempt moving either of them.

I think this bug was likely present in the original pull request too --
although it was moving the selection rather than moving items within the
list -- but I didn't notice it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>