Fix memoization
Commit 5621c4e21854 fixed a terrible bug that affected the correctness
of the matching algorithm, but it did so at the cost of the speed of our
pathological test case (although other test cases seemed slightly
better).

  Buggy version:
                          user     system      total        real
  pathological        0.000000   0.000000   0.000000 (  0.003640)
  command-t           0.390000   0.000000   0.390000 (  0.383473)
  chromium (subset)   1.170000   0.010000   1.180000 (  0.525694)
  chromium (whole)    4.550000   0.010000   4.560000 (  1.579863)

  Fixed version:
                          user     system      total        real
  pathological        1.880000   0.010000   1.890000 (  1.887996)
  command-t           0.350000   0.000000   0.350000 (  0.348788)
  chromium (subset)   1.140000   0.010000   1.150000 (  0.504924)
  chromium (whole)    4.430000   0.000000   4.430000 (  1.516979)

After fixing this, I went back and audited the memoization code to see
if we could "have our cake and eat it too"; that is, get robust/correct
matching and keep some or all of the speed gains that we got from
memoization in the buggy version.

It turns out, we can:

                          user     system      total        real
  pathological        0.000000   0.000000   0.000000 (  0.002413)
  command-t           0.380000   0.000000   0.380000 (  0.380633)
  chromium (subset)   0.850000   0.010000   0.860000 (  0.441269)
  chromium (whole)    3.400000   0.010000   3.410000 (  1.342450)

So, we got faster on every test case in the benchmark suite. Not sure
what I was smoking when I originally wrote the memoization code; instead
of memoizing a known result at the earliest point in the haystack
string, I was doing so at the latest, which reduces the value of the
memoization.

Signed-off-by: Greg Hurrell <greg@hurrell.net>