Update installation docs
De-emphasize Vimball as an installation method (I consider it a broken
format, not just because of the way it gloms all plug-ins into a single
shared namespace, but because it frequently fails to extract correctly,
leading to support tickets). I'd like to eventually even stop
distributing the Vimball at all, but the first gentle step is not
documenting Vimball's existence.

Add info about some of the more popular plug-in managers. Pathogen is
still king of simplicity and robustness, but the allure of shiny,
overengineered things and the widespread misunderstanding of Git
submodules has caused the marketplace to fragment. Oh well.