Consider 'autowriteall' when determining default open command
Like `'hidden'`, `'autowriteall'` makes it safe to abandon buffers with
unsaved changes, so we can issue a standard `:e` command in this case
and let Vim save modified files for us beforehand.

Note that we have to include a check to ensure that `'readonly'` isn't
set, because that would prevent `'autowriteall'` from doing its thing.

Fixes:

  https://github.com/wincent/Command-T/issues/88