Tweak word clearing to match shell behavior
The word clearing algorithm was clearing either words or whitespace. For
example, given the following (where ^ indicates the cursor position):

  foo      bar     baz
                      ^

Hitting <C-w> three times would first delete "baz", then the whitespace
after "bar", finally "bar" itself.

This commit changes the behavior to match that of the shell:

  foo bar
              ^ (<C-w> deletes trailing whitespace and "bar")
  foo       bar
        ^ (<C-w> deletes "foo" and whitespace to left of cursor)
  foobar
     ^ (<C-w> deletes "foo")
  foo
  ^ (<C-W> does nothing)

For clarity and simplicity, switched to a largely regex-based approach
which has no conditionals in it and is probably more readable.