Abbreviate license headers
There are plenty of diverse opinions on whether it's worth it to include
full license text in every file:

- http://lu.is/blog/2012/03/17/on-the-importance-of-per-file-license-information/
- http://softwarefreedom.org/resources/2012/ManagingCopyrightInformation.html
- http://stackoverflow.com/questions/845895/putting-license-in-each-code-file

The main argument against is verbosity. The main argument for is
ensuring that licensing info stays with the file even if it gets pulled
out of a project.

Seems that for a project like this one, the cost of license abuse is
low, and people's awareness of what the "BSD 2-clause license" is is
relatively good, so a terse point to the BSD license should suffice.