Disallow scrolling
We never intended to allow scrolling (see the previous commit) but it
was actually possible to make it happen with some careful staging.

Namely, you could trick Command-T into scrolling by constraining the
amount of available space by opening many splits; this didn't affect the
`#max_lines` calculation, so Command-T thought it has more room than it
actually did. Under these circumstances, you'd hit the end of the range
and be able to scroll past it to see more matches; not broken, but not
intended either.

Compare that with what happens when you change the size of your terminal
window; _that_ constraint does get picked up.

With this change, we set the desired window height and then read it back
to see what we actually got, effectively preventing scrolling because we
never put more results in the buffer than would fit on the screen.