Further robustify scroll-prevention
The old version of this code worked by inverting the order of the match
list results and then allowing the rest of the code to proceed as usual.
Now that we're trying to deal with the nuances of things like different
constraint models (external resizing vs internal pressure from splits)
things will be more robust if we recognize that the match results may be
more than we can display, and explicitly rendering things in reverse.

This gives us some nice benefits, such as better handling edge cases
like having `g:CommandTMatchWindowReverse` set in conjunction with
`g:CommandTMinHeight` and too few results to fill the height; previously
we'd show them up the top of the results window instead of the bottom,
but now we do the right thing.

Overall, the code gets simpler in some places and more complicated in
others. It would be nice to follow up later with a refactoring to clean
up this class.