Simplify require statements
No need to pull in "command-t/vim" as that gets pulled in from
"command-t/scanner/file_scanner".

Instead of "command-t/scanner/file_scanner", pull in
"command-t/scanner/file_scanner/find_file_scanner", which is our
superclass; as it was, we had an ordering dependency which relied on
some other file having required the superclass before we get evaluated,
which isn't good.

There's no need to pull in "open3" as the superclass is already doing
that.

Explicit is better than implicit and all that, but this is Ruby, land of
magic, and of slow require statements.