doc: improve trouble-shooting docs
Based on discussion in https://github.com/wincent/Command-T/issues/97
(in which a user had an old version of Command-T from a previous vimball
install, in addition to the latest version that he was trying to install
via a plugin manager), add a tip about removing old versions.

Also, try to future-proof the text with OS X recommendations (which
unfortunately have to change with every OS X release).