Fix ignoring of 'wildignore'
Since 0b8c6b44142fbb we've been ignoring Vim's built-in `'wildignore'`
setting and only filtering if `g:CommandTWildIgnore` is set.

We need to distinguish between these cases:

- user leaves `g:CommandTWildIgnore` unset; we fall through to
  `'wildignore'`, which defaults to "", and if it is empty, we do no
  filtering; otherwise we do filtering
- user sets `g:CommandTWildIgnore`; this overrides `'wildignore'` and we
  do filtering whenever it is non-empty

I have a sense that the handling might be able to be simplified, but for
now I just want to get the fix in.

Fixes:

  https://github.com/wincent/Command-T/issues/106