Consistently use parens around method param lists
There was a mix of styles scattered throughout the codebase:

  def bar boom, bang
    # ...
  end

  def foo(bar, baz, bing)
    # ...
  end

Generally the older methods use the former style and the newer ones use
the latter. The early tendency to eschew params was driven by a desire
to embrace the "Zen of Ruby", which seemed to want to minimize use of
punctuation.

Since then, however, I've come around. The real "Zen" is actually to
minimize the use of punctuation but never to the extent that it reduces
readability. This is why Ruby has postfix conditionals and doesn't
require semi-colons to terminate statements (reduce clutter, improve
flow) but it has plenty of syntax involving punctuation.

In practice, the parens in function parameter lists serve as a useful
and instantly recognizable visual cue. They help readability more than
they introduce clutter.

Having worked on teams now with many non-Ruby programmers coming to
Ruby, I saw the confusion that paren-less method declarations brought.
And a quick survey of some high-profile Ruby projects shows that they
almost universally use parens in their param lists.